+++ configure.py
@@ -1439,6 +1439,7 @@ alternator = [
        'alternator/server.cc',
        'alternator/executor.cc',
        'alternator/executor_read.cc',
+       'alternator/executor_util.cc',
        'alternator/stats.cc',
        'alternator/serialization.cc',
        'alternator/expressions.cc',
